Mega cleanup (read commit message)
* Breaking API change: Blocks and items fields are no longer available in API package, use Item.REGISTRY and Block.REGISTRY instead
* Breaking API change: some fields and methods in FrogAPI are removed
* Fix FrogFuelHandler
* FrogAchievements.get now return nonnull; failure on reflection now throws exception
* Auto-cleanup done by Intellij IDEA
* No new content was introduced
* Use mcmod.info to handle dependencies

+++ src/main/java/frogcraftrebirth/client/gui/GuiAirPump.java
@@ -31,10 +31,10 @@ public class GuiAirPump extends GuiTileFrog<TileAirPump, ContainerAirPump> {
 	@Override
 	protected void drawGuiContainerBackgroundLayer(float partialTicks, int mouseX, int mouseY) {
 		super.drawGuiContainerBackgroundLayer(partialTicks, mouseX, mouseY);
-		int chargeIcon = (int) (14 * tile.charge / 10000); // TileAirPump.MAX_CHARGE = 10000;
+		int chargeIcon = 14 * tile.charge / 10000; // TileAirPump.MAX_CHARGE = 10000;
 		this.drawTexturedModalRect(this.guiLeft + 118, this.guiTop + 33 + 14 - chargeIcon, 188, 14 - chargeIcon, 10, chargeIcon);
 		
-		int airPercentage = (int) (40 * tile.airAmount() / 1000); //TileAirPump.MAX_AIR = 1000;
+		int airPercentage = 40 * tile.airAmount() / 1000; //TileAirPump.MAX_AIR = 1000;
 		this.drawTexturedModalRect(this.guiLeft + 145, this.guiTop + 63 - airPercentage, 176, 0, 12, airPercentage);
 		this.drawTexturedModalRect(this.guiLeft + 145, this.guiTop + 60 - airPercentage, 176, 41, 12, 4);
 	}
