+++ src/main/java/frogcraftrebirth/client/gui/GuiPyrolyzer.java
@@ -38,11 +38,11 @@ public class GuiPyrolyzer extends GuiTileFrog<TilePyrolyzer, ContainerPyrolyzer>
 	@Override
 	protected void drawGuiContainerBackgroundLayer(float partialTicks, int mouseX, int mouseY) {
 		super.drawGuiContainerBackgroundLayer(partialTicks, mouseX, mouseY);
-		int chargeIcon = (int) (14 * tile.charge / tile.maxCharge);
+		int chargeIcon = 14 * tile.charge / tile.maxCharge;
 		this.drawTexturedModalRect(this.guiLeft + 81, this.guiTop + 57 + 14 - chargeIcon, 176, 52 + 14 - chargeIcon, 10, chargeIcon);
 		if (tile.isWorking()) {
 			this.drawTexturedModalRect(this.guiLeft + 25, this.guiTop + 49, 176, 66, 14, 14);
-			int progressPercent = tile.processMax == 0 ? 0 : (int) (24 * tile.process / tile.processMax);
+			int progressPercent = tile.processMax == 0 ? 0 : 24 * tile.process / tile.processMax;
 			this.drawTexturedModalRect(this.guiLeft + 45, this.guiTop + 29, 176, 80, progressPercent, 17);
 		}
 		this.renderFluidTank(tile.tank, this.guiLeft + 143, this.guiTop + 23, 16, 47);
